feat: improve args ergonomics; fix dupe handling and validation args
Provide explanation of the dupe handling in the project config examples
  `config.example.ini`
  `configs/dot-ilma.conf.example`
and the README and fix the naming pattern to:
  `ilma project/`    # first pass
  > project.bak
  `ilma project/`    # second pass
  > project.1.bak
Timestamps work similarly.

Add multi-backup mode and create arch short forms.

  -a, --archive
  -b, --backup
  -c, --context
  -e, --encrypt
  -r, --remote

When a user runs `ilma -acb`, ilma will creat all three of and archive,
context mirror, and backup.

Errata.
Previous fixes to the code regressed:
  - `ilma validate --compression`
  - `ilma validate --dependencies`
  .. `--gpg`
This has been fixed.